R administration page displays R server clusters
