+++ opal-gwt-client/src/main/java/org/obiba/opal/web/gwt/app/client/administration/r/RAdministrationView.ui.xml
@@ -25,121 +25,15 @@
       <ui:msg description="R servers admin label">R Servers</ui:msg>
     </b:Heading>
 
-    <o:NavTabsPanel ui:field="tabs" addStyleNames="top-margin">
-      <g:Anchor>
-        <ui:msg description="R server status label">Status</ui:msg>
-      </g:Anchor>
-      <g:FlowPanel>
-        <b:Paragraph>
-          <ui:msg description="R server text">
-            Whether the R server can be started and stopped depends on the configuration of the system. Stopping the R
-            server will destroy all current client connections.
-          </ui:msg>
-        </b:Paragraph>
-        <g:FlowPanel addStyleNames="large-bottom-margin">
-          <b:ButtonGroup>
-            <b:Button ui:field="startStopButton" type="INVERSE">
-              Start/Stop
-            </b:Button>
-          </b:ButtonGroup>
-          <b:ButtonGroup>
-            <b:Button ui:field="rTestButton" type="SUCCESS">
-              <ui:msg description="R server test label">Test connection with R</ui:msg>
-            </b:Button>
-          </b:ButtonGroup>
-        </g:FlowPanel>
-        <g:FlowPanel addStyleNames="xlarge-bottom-margin">
-          <b:Paragraph>
-            <ui:msg description="Rserve logs description">R server activity logs can be downloaded in their raw form.</ui:msg>
-          </b:Paragraph>
-          <b:Button ui:field="downloadLogs" icon="DOWNLOAD" type="INFO">
-            <ui:msg description="Download Rserve logs label">Download</ui:msg>
-          </b:Button>
-        </g:FlowPanel>
+    <b:Paragraph>
+      <ui:msg description="R Servers information">
+        R servers are grouped by clusters. In each cluster all the R servers are considered to be interchangeable, i.e.
+        a load balancer can select one or the other on R session creation. On the other hand, R servers from different
+        clusters are expected to be different (packages installed, DataSHIELD settings etc.).
+      </ui:msg>
+    </b:Paragraph>
 
-      </g:FlowPanel>
-
-      <g:Anchor>
-        <ui:msg description="R server instances label">Servers</ui:msg>
-      </g:Anchor>
-
-      <g:FlowPanel>
-
-        <b:Paragraph>
-          <ui:msg description="R servers text">
-            List of all the R servers in the cluster.
-          </ui:msg>
-        </b:Paragraph>
-
-        <g:FlowPanel>
-
-        </g:FlowPanel>
-
-        <g:FlowPanel>
-          <g:FlowPanel addStyleNames="pull-left">
-            <b:ButtonGroup>
-              <b:Button ui:field="manageRServers" icon="PENCIL" type="INFO">
-                <ui:msg description="R servers management label">Manage</ui:msg>
-              </b:Button>
-            </b:ButtonGroup>
-            <b:ButtonGroup>
-              <b:Button ui:field="refreshServers" icon="REFRESH">
-                <ui:msg description="Refresh R packages label">Refresh</ui:msg>
-              </b:Button>
-            </b:ButtonGroup>
-          </g:FlowPanel>
-          <o:OpalSimplePager ui:field="serversPager" location="RIGHT" addStyleNames="small-indent pull-right"/>
-          <b:Controls addStyleNames="pull-right small-indent">
-            <b:InputAddOn prependIcon="FILTER">
-              <o:TextBoxClearable ui:field="serversFilter" addStyleNames="inline-block"/>
-            </b:InputAddOn>
-          </b:Controls>
-        </g:FlowPanel>
-        <o:Table ui:field="serversTable" addStyleNames="small-top-margin" pageSize="20"/>
-
-      </g:FlowPanel>
-
-      <g:Anchor>
-        <ui:msg description="R server packages label">Packages</ui:msg>
-      </g:Anchor>
-      <g:FlowPanel>
-
-        <b:Paragraph>
-          <ui:msg description="R server packages text">
-            List of all the R packages installed on the R servers. Removing or installing an R package applies to all R servers.
-          </ui:msg>
-        </b:Paragraph>
-
-        <g:FlowPanel>
-          <g:FlowPanel addStyleNames="pull-left">
-            <b:ButtonGroup>
-              <b:DropdownButton text="Install" icon="PLUS" type="INFO">
-                <b:NavLink ui:field="installPackage">
-                  <ui:msg description="Server logs label">Install package...</ui:msg>
-                </b:NavLink>
-                <b:NavLink ui:field="updateAllPackages">
-                  <ui:msg description="Update R packages label">Update all packages...</ui:msg>
-                </b:NavLink>
-              </b:DropdownButton>
-            </b:ButtonGroup>
-            <b:ButtonGroup>
-              <b:Button ui:field="refreshPackages" icon="REFRESH">
-                <ui:msg description="Refresh R packages label">Refresh</ui:msg>
-              </b:Button>
-            </b:ButtonGroup>
-          </g:FlowPanel>
-          <o:OpalSimplePager ui:field="packagesPager" location="RIGHT" addStyleNames="small-indent pull-right"/>
-          <b:Controls addStyleNames="pull-right small-indent">
-            <b:InputAddOn prependIcon="FILTER">
-              <o:TextBoxClearable ui:field="packagesFilter" addStyleNames="inline-block"/>
-            </b:InputAddOn>
-          </b:Controls>
-        </g:FlowPanel>
-        <o:Table ui:field="packagesTable" addStyleNames="small-top-margin" pageSize="20"/>
-
-      </g:FlowPanel>
-
-    </o:NavTabsPanel>
+    <b:TabPanel tabPosition="left" ui:field="clusterTabs"/>
 
     <g:FlowPanel>
       <b:Heading size="4">
