+++ opal-gwt-client/src/main/java/org/obiba/opal/web/gwt/app/client/administration/r/list/RClusterView.ui.xml
@@ -1,0 +1,137 @@
+<!--
+  ~ Copyright (c) 2021 OBiBa. All rights reserved.
+  ~
+  ~ This program and the accompanying materials
+  ~ are made available under the terms of the GNU Public License v3.0.
+  ~
+  ~ You should have received a copy of the GNU General Public License
+  ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  -->
+
+<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
+    xmlns:g="urn:import:com.google.gwt.user.client.ui"
+    xmlns:o="urn:import:org.obiba.opal.web.gwt.app.client.ui"
+    xmlns:b="urn:import:com.github.gwtbootstrap.client.ui"
+    ui:generateFormat='com.google.gwt.i18n.rebind.format.PropertiesFormat'
+    ui:generateKeys="com.google.gwt.i18n.server.keygen.MD5KeyGenerator"
+    ui:generateLocales="default">
+
+  <g:FlowPanel>
+
+    <g:FlowPanel>
+      <b:Paragraph>
+        <ui:msg description="R server text">
+          Whether the R server can be started and stopped depends on the configuration of the system. Stopping the R
+          server will destroy all current client connections.
+        </ui:msg>
+      </b:Paragraph>
+      <g:FlowPanel addStyleNames="large-bottom-margin">
+        <b:ButtonGroup>
+          <b:Button ui:field="startStopButton" type="INVERSE">
+            Start/Stop
+          </b:Button>
+        </b:ButtonGroup>
+        <b:ButtonGroup>
+          <b:Button ui:field="rTestButton" type="SUCCESS">
+            <ui:msg description="R server test label">Test connection with R</ui:msg>
+          </b:Button>
+        </b:ButtonGroup>
+      </g:FlowPanel>
+      <g:FlowPanel addStyleNames="large-bottom-margin">
+        <b:Paragraph>
+          <ui:msg description="Rserve logs description">R server activity logs can be downloaded in their raw form.</ui:msg>
+        </b:Paragraph>
+        <b:Button ui:field="downloadLogs" icon="DOWNLOAD" type="INFO">
+          <ui:msg description="Download Rserve logs label">Download</ui:msg>
+        </b:Button>
+      </g:FlowPanel>
+
+    </g:FlowPanel>
+
+    <o:NavTabsPanel ui:field="tabs" addStyleNames="top-margin">
+
+      <g:Anchor>
+        <ui:msg description="R server instances label">Servers</ui:msg>
+      </g:Anchor>
+
+      <g:FlowPanel>
+
+        <b:Paragraph>
+          <ui:msg description="R servers text">
+            List of all the R servers in the cluster.
+          </ui:msg>
+        </b:Paragraph>
+
+        <g:FlowPanel>
+
+        </g:FlowPanel>
+
+        <g:FlowPanel>
+          <g:FlowPanel addStyleNames="pull-left">
+            <b:ButtonGroup>
+              <b:Button ui:field="manageRServers" icon="PENCIL" type="INFO">
+                <ui:msg description="R servers management label">Manage</ui:msg>
+              </b:Button>
+            </b:ButtonGroup>
+            <b:ButtonGroup>
+              <b:Button ui:field="refreshServers" icon="REFRESH">
+                <ui:msg description="Refresh R packages label">Refresh</ui:msg>
+              </b:Button>
+            </b:ButtonGroup>
+          </g:FlowPanel>
+          <o:OpalSimplePager ui:field="serversPager" location="RIGHT" addStyleNames="small-indent pull-right"/>
+          <b:Controls addStyleNames="pull-right small-indent">
+            <b:InputAddOn prependIcon="FILTER">
+              <o:TextBoxClearable ui:field="serversFilter" addStyleNames="inline-block"/>
+            </b:InputAddOn>
+          </b:Controls>
+        </g:FlowPanel>
+        <o:Table ui:field="serversTable" addStyleNames="small-top-margin" pageSize="20"/>
+
+      </g:FlowPanel>
+
+      <g:Anchor>
+        <ui:msg description="R server packages label">Packages</ui:msg>
+      </g:Anchor>
+      <g:FlowPanel>
+
+        <b:Paragraph>
+          <ui:msg description="R server packages text">
+            List of all the R packages installed on the R servers. Removing or installing an R package applies to all R servers.
+          </ui:msg>
+        </b:Paragraph>
+
+        <g:FlowPanel>
+          <g:FlowPanel addStyleNames="pull-left">
+            <b:ButtonGroup>
+              <b:DropdownButton text="Install" icon="PLUS" type="INFO">
+                <b:NavLink ui:field="installPackage">
+                  <ui:msg description="Server logs label">Install package...</ui:msg>
+                </b:NavLink>
+                <b:NavLink ui:field="updateAllPackages">
+                  <ui:msg description="Update R packages label">Update all packages...</ui:msg>
+                </b:NavLink>
+              </b:DropdownButton>
+            </b:ButtonGroup>
+            <b:ButtonGroup>
+              <b:Button ui:field="refreshPackages" icon="REFRESH">
+                <ui:msg description="Refresh R packages label">Refresh</ui:msg>
+              </b:Button>
+            </b:ButtonGroup>
+          </g:FlowPanel>
+          <o:OpalSimplePager ui:field="packagesPager" location="RIGHT" addStyleNames="small-indent pull-right"/>
+          <b:Controls addStyleNames="pull-right small-indent">
+            <b:InputAddOn prependIcon="FILTER">
+              <o:TextBoxClearable ui:field="packagesFilter" addStyleNames="inline-block"/>
+            </b:InputAddOn>
+          </b:Controls>
+        </g:FlowPanel>
+        <o:Table ui:field="packagesTable" addStyleNames="small-top-margin" pageSize="20"/>
+
+      </g:FlowPanel>
+
+    </o:NavTabsPanel>
+
+  </g:FlowPanel>
+
+</ui:UiBinder>
