+++ opal-r/src/main/java/org/obiba/opal/web/r/RServiceClustersResource.java
@@ -12,14 +12,13 @@ package org.obiba.opal.web.r;
 
 import org.obiba.opal.r.service.RServerManagerService;
 import org.obiba.opal.web.model.OpalR;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -28,14 +27,14 @@ import java.util.stream.Collectors;
 @Path("/service/r/clusters")
 public class RServiceClustersResource {
 
-  private static final Logger log = LoggerFactory.getLogger(RServiceClustersResource.class);
-
   @Autowired
   private RServerManagerService rServerManagerService;
 
   @GET
   public List<OpalR.RServerClusterDto> getClusters() {
     return rServerManagerService.getRServerClusters().stream()
-        .map(Dtos::asDto).collect(Collectors.toList());
+        .map(Dtos::asDto)
+        .sorted(Comparator.comparing(OpalR.RServerClusterDto::getName))
+        .collect(Collectors.toList());
   }
 }
